# Use cache for fetchObject